+++ packages/authhero/src/constants/index.ts
@@ -1,0 +1,2 @@
+export { DEFAULT_THEME } from "./defaultTheme";
+export { DEFAULT_BRANDING } from "./defaultBranding";